Move subscription model into zuora-core
