+++ handlers/holiday-stop-api/src/main/scala/com/gu/holiday_stops/PotentialHolidayStops.scala
@@ -2,7 +2,7 @@ package com.gu.holiday_stops
 
 import java.time.LocalDate
 
-import com.gu.holiday_stops.subscription.HolidayStopCredit
+import com.gu.zuora.subscription.HolidayStopCredit
 import play.api.libs.json.Reads._
 import play.api.libs.json._
