+++ lib/holiday-stops/src/main/scala/com/gu/holiday_stops/Config.scala
@@ -3,6 +3,7 @@ package com.gu.holiday_stops
 import com.gu.salesforce.SFAuthConfig
 import com.gu.util.config.LoadConfigModule.StringFromS3
 import com.gu.util.config.{ConfigLocation, LoadConfigModule, Stage}
+import com.gu.zuora.subscription.{CreditProduct, OverallFailure}
 import play.api.libs.json.{Json, Reads}
 
 case class Config(
