+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/Fixtures.scala
@@ -2,11 +2,11 @@ package com.gu.holiday_stops
 
 import java.time.LocalDate
 
-import com.gu.holiday_stops.subscription.{RatePlan, RatePlanCharge, Subscription}
-import com.gu.salesforce.RecordsWrapperCaseClass
-import com.gu.salesforce.SFAuthConfig
+import com.gu.salesforce.{RecordsWrapperCaseClass, SFAuthConfig}
 import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequest.{HolidayStopRequest, HolidayStopRequestActionedCount, HolidayStopRequestEndDate, HolidayStopRequestIsWithdrawn, HolidayStopRequestStartDate}
-import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequestsDetail._
+import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequestsDetail.{HolidayStopRequestId, HolidayStopRequestsDetail, HolidayStopRequestsDetailChargePrice, HolidayStopRequestsDetailExpectedInvoiceDate, HolidayStopRequestsDetailId, ProductName}
+import com.gu.zuora.subscription.Fixtures.mkRatePlanCharge
+import com.gu.zuora.subscription._
 import io.circe.generic.auto._
 import io.circe.parser.decode
 import org.scalatest.Assertions
@@ -22,70 +22,6 @@ object Fixtures extends Assertions {
     case "Semi_Annual" => 6
     case "Specific_Weeks" => 1
   }
-
-  def mkRatePlanCharge(
-    name: String,
-    price: Double,
-    billingPeriod: String,
-    chargedThroughDate: Option[LocalDate] = Some(LocalDate.of(2019, 9, 2)),
-    effectiveStartDate: LocalDate = LocalDate.of(2019, 6, 2),
-    specificBillingPeriod: Option[Int] = None,
-    upToPeriodsType: Option[String] = None,
-    upToPeriods: Option[Int] = None,
-    endDateCondition: Option[String] = Some("Subscription_End"),
-  ) = RatePlanCharge(
-    name = name,
-    number = "C1",
-    price,
-    Some(billingPeriod),
-    effectiveStartDate,
-    chargedThroughDate,
-    HolidayStart__c = None,
-    HolidayEnd__c = None,
-    processedThroughDate = chargedThroughDate.map(_.minusMonths(billingPeriodToMonths(billingPeriod))),
-    productRatePlanChargeId = "",
-    specificBillingPeriod = specificBillingPeriod,
-    endDateCondition = endDateCondition,
-    upToPeriodsType = upToPeriodsType,
-    upToPeriods = upToPeriods
-  )
-
-  def mkGuardianWeeklySubscription(
-    termStartDate: LocalDate = LocalDate.now(),
-    termEndDate: LocalDate = LocalDate.now(),
-    customerAcceptanceDate: LocalDate = LocalDate.now(),
-    price: Double = -1.0,
-    billingPeriod: String = "Quarter",
-    chargedThroughDate: Option[LocalDate] = None,
-    effectiveStartDate: LocalDate = LocalDate.now()
-  ): Subscription =
-    Subscription(
-      subscriptionNumber = "S1",
-      termStartDate,
-      termEndDate,
-      customerAcceptanceDate,
-      currentTerm = 12,
-      currentTermPeriodType = "Month",
-      autoRenew = true,
-      ratePlans = List(
-        RatePlan(
-          productName = "Guardian Weekly - Domestic",
-          ratePlanName = "GW Oct 18 - Quarterly - Domestic",
-          ratePlanCharges =
-            List(mkRatePlanCharge(
-              "GW Oct 18 - Quarterly - Domestic",
-              price,
-              billingPeriod,
-              chargedThroughDate,
-              effectiveStartDate
-            )),
-          productRatePlanId = "",
-          id = "",
-          lastChangeType = None
-        )
-      ),
-      status = "Active"
-    )
 
   def mkSubscriptionWithHolidayStops() = Subscription(
     status = "Active",
@@ -302,7 +238,7 @@ object Fixtures extends Assertions {
       Product_Name__c = ProductName(productName),
       Stopped_Publication_Date__c = AffectedPublicationDate(stopDate),
       Estimated_Price__c = estimatedPrice.map(HolidayStopRequestsDetailChargePrice.apply),
-      Charge_Code__c = chargeCode.map(RatePlanChargeCode),
+      Charge_Code__c = chargeCode.map(RatePlanChargeCode(_)),
       Actual_Price__c = actualPrice.map(HolidayStopRequestsDetailChargePrice.apply),
       Expected_Invoice_Date__c = expectedInvoiceDate.map(HolidayStopRequestsDetailExpectedInvoiceDate.apply)
     )
