+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/SubscriptionTest.scala
@@ -1,8 +1,8 @@
-package com.gu.holiday_stops.subscription
+package com.gu.holiday_stops
 
 import java.time.LocalDate
 
-import com.gu.holiday_stops.Fixtures
+import com.gu.zuora.subscription.{RatePlanCharge, Fixtures => SubscriptionFixtures}
 import org.scalactic.TypeCheckedTripleEquals
 import org.scalatest.{FlatSpec, Matchers, OptionValues}
 
@@ -57,7 +57,7 @@ class SubscriptionTest extends FlatSpec with Matchers with OptionValues with Typ
   }
 
   it should "give no ratePlanCharge when subscription has no holiday stops applied" in {
-    val subscription = Fixtures.mkGuardianWeeklySubscription(
+    val subscription = SubscriptionFixtures.mkGuardianWeeklySubscription(
       termStartDate = LocalDate.of(2018, 1, 1),
       termEndDate = LocalDate.of(2019, 1, 1),
       price = 123,
