+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/ZuoraSttpEffects.scala
@@ -1,6 +1,6 @@
 package com.gu.holiday_stops
 
-import com.gu.holiday_stops.subscription.Subscription
+import com.gu.zuora.subscription.Subscription
 import com.softwaremill.sttp.testing.SttpBackendStub
 import com.softwaremill.sttp.{Id, Response}
