+++ lib/holiday-stops/src/test/scala/com/gu/salesforce/holiday_stops/SalesforceHolidayStopRequestEndToEndEffectsTest.scala
@@ -3,15 +3,14 @@ package com.gu.salesforce.holiday_stops
 import java.time.LocalDate
 
 import com.gu.effects.{GetFromS3, RawEffects}
-import com.gu.holiday_stops.Fixtures
-import com.gu.holiday_stops.subscription.{Subscription, SubscriptionData}
-import com.gu.salesforce.{IdentityId, SFAuthConfig, SalesforceClient}
 import com.gu.salesforce.SalesforceReads._
 import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequest._
 import com.gu.salesforce.holiday_stops.SalesforceHolidayStopRequestsDetail._
+import com.gu.salesforce.{IdentityId, SFAuthConfig, SalesforceClient}
 import com.gu.test.EffectsTest
 import com.gu.util.config.{LoadConfigModule, Stage}
 import com.gu.util.resthttp.JsonHttp
+import com.gu.zuora.subscription._
 import org.scalatest.{FlatSpec, Matchers}
 import scalaz.{-\/, \/-}
