+++ lib/zuora-core/src/main/scala/com/gu/zuora/subscription/HolidayStopCredit.scala
@@ -1,4 +1,4 @@
-package com.gu.holiday_stops.subscription
+package com.gu.zuora.subscription
 
 import java.time.LocalDate
