+++ lib/zuora-core/src/main/scala/com/gu/zuora/subscription/package.scala
@@ -1,6 +1,6 @@
-package com.gu
+package com.gu.zuora
 
-package object holiday_stops {
+package object subscription {
   type ZuoraApiResponse[T] = Either[ZuoraApiFailure, T]
   type SalesforceApiResponse[T] = Either[SalesforceApiFailure, T]
 }
