+++ lib/zuora-core/src/test/scala/com/gu/zuora/subscription/SubscriptionDataGuardianWeeklyWithRemovedRatePlansIntegrationTest.scala
@@ -1,10 +1,10 @@
-package com.gu.holiday_stops.subscription
+package com.gu.zuora.subscription
 
 import java.time.DayOfWeek.FRIDAY
 import java.time.LocalDate
 
-import com.gu.holiday_stops.subscription.SubscriptionDataIntegrationTest.testSubscriptonDataIssueGeneration
 import com.gu.zuora.ZuoraProductTypes
+import com.gu.zuora.subscription.SubscriptionDataIntegrationTest.testSubscriptonDataIssueGeneration
 import org.scalatest.FlatSpec
 
 class SubscriptionDataGuardianWeeklyWithRemovedRatePlansIntegrationTest extends FlatSpec {
